Don't fall back to global default if the project specifies no default assignee
Previously, the behaviour was to use the assignee specified in the configuration
screen if the project had no default assignee set. I don't think this makes
sense; you may want to be able to set a blank assignee in certain projects, and
it's unlikely you'll have so many projects that setting an assignee on each one
will be that difficult.